pyproject.toml
[tool.poetry]
name = "ctk_api"
version = "0.1.0"
description = "A beautiful description."
authors = ["Reinder Vos de Wael <reinder.vosdewael@childmind.org>"]
license = "LGPL-2.1"
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
pydantic = "^2.6.3"
python-docx = "^1.1.0"
pydantic-settings = "^2.2.1"
python-multipart = "^0.0.9"
uvicorn = "^0.27.1"
openai = "^1.12.0"
pypandoc = "^1.13"
pyyaml = "6.0"
email-validator = "^2.1.1"
sqlalchemy = "^2.0.27"
psycopg2 = "^2.9.9"
polars = "^0.20.14"
pytz = "^2024.1"
docxcompose = "^1.4.0"
spacy = "^3.7.4"
mlconjug3 = "^3.11.0"
pyjwt = {extras = ["crypto"], version = "^2.8.0"}

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.2"
mypy = "^1.8.0"
pre-commit = "^3.6.2"
pytest-cov = "^4.1.0"
httpx = "^0.27.0"
pytest-dotenv = "^0.5.2"
pytest-mock = "^3.12.0"
ruff = "^0.3.0"
types-pyyaml = "^6.0.12.12"
pytest-asyncio = "^0.23.2"
types-pytz = "^2024.1.0.20240203"
types-python-dateutil = "^2.9.0.20240316"

[tool.poetry.group.docs.dependencies]
pdoc = "^14.4.0"

[tool.poetry.extras]
docs = ["pdoc"]

[tool.pytest.ini_options]
pythonpath = [
  "src"
]
env_files = [
  "./tests/.test.env"
]

[tool.mypy]
ignore_missing_imports = true

[tool.ruff]
exclude = [
  ".bzr",
  ".direnv",
  ".eggs",
  ".git",
  ".git-rewrite",
  ".hg",
  ".mypy_cache",
  ".nox",
  ".pants.d",
  ".pytype",
  ".ruff_cache",
  ".svn",
  ".tox",
  ".venv",
  "__pypackages__",
  "_build",
  "buck-out",
  "build",
  "dist",
  "node_modules",
  "venv"
]
line-length = 88
indent-width = 4
src = ["src"]
target-version = "py311"

[tool.ruff.lint]
select = ["ALL"]
ignore = [
  "ANN101",  # Self should never be type annotated.
  "ANN102",  # cls should never be type annotated.
  "B008"  # Allow function call in arguments; this is common in FastAPI.
]
fixable = ["ALL"]
unfixable = []
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
  "S101",  # Allow assets
  "ARG",  # Unused arguments are common in tests (fixtures).
  "FBT",  # Allow positional booleans
  "SLF001",  # Allow private member access.
  "INP001"  # No need for namespace packages in tests.
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[build-system]
requires = ["poetry-core>=1.2.0"]
build-backend = "poetry.core.masonry.api"
